% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggradar-package.R
\docType{package}
\name{ggradar2-package}
\alias{ggradar2-package}
\title{ggradar2: Create radar charts using ggplot2}
\description{
This package exports the functionality of ggplot2 to create radar charts.
}
\author{
\strong{Maintainer}: Ricardo Bion \email{ricardo.bion@airbnb.com}

Authors:
\itemize{
  \item Pedro Sartori [contributor]
}

Other contributors:
\itemize{
  \item Weiping Mei \email{meiweipingg@163.com} [contributor]
}

}
\keyword{internal}
