% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggradar2.R
\name{ggradar2}
\alias{ggradar2}
\title{ggradar2}
\usage{
ggradar2(plot.data, base.size = 20, webtype = "mini",
  axis.labels = "", grid.min = 0, grid.max = 1, centre.y = grid.min
  - ((1/9) * (grid.max - grid.min)), label.centre.y = FALSE,
  grid.line.width = 0.5, grid.line.trend = "classic",
  gridline.min.linetype = "longdash",
  gridline.mid.linetype = "longdash",
  gridline.max.linetype = "longdash", gridline.min.colour = "grey",
  gridline.mid.colour = "#007A87", gridline.max.colour = "grey",
  grid.label.size = 6, gridline.label.offset = -0.1 * (grid.max -
  centre.y), label.gridline.min = TRUE, label.gridline.mid = TRUE,
  label.gridline.max = TRUE, gridline.label = NULL,
  axis.label.offset = 1.15, axis.label.size = 5,
  axis.line.colour = "grey", group.line.width = 1.5,
  group.point.size = 3, group.colours = NULL,
  group.fill.colours = NULL, background.circle.colour = "#D7D6D1",
  background.circle.transparency = 0.2, legend.title = "",
  plot.legend = TRUE, plot.title = "", legend.text.size = 14,
  radarshape = "round", polygonfill = TRUE,
  polygonfill.transparency = 0.2, multiplots = "none",
  stripbackground = TRUE, fullscore = NULL)
}
\description{
ggradar2
}
\author{
Liang Xu
}
