% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggradar2.R
\name{ggradar2}
\alias{ggradar2}
\title{ggradar2}
\usage{
ggradar2(
  plot.data,
  base.size = 15,
  font.radar = "sans",
  values.radar = c("0\%", "20\%", "40\%", "60\%", "80\%", "100\%"),
  axis.labels = colnames(plot.data)[-1],
  grid.min = 0,
  grid.n2 = 0.2,
  grid.n3 = 0.4,
  grid.n4 = 0.6,
  grid.n5 = 0.8,
  grid.max = 1,
  gridlines.show = c(TRUE, TRUE, TRUE, TRUE, TRUE, TRUE),
  centre.y = grid.min - ((1/9) * (grid.max - grid.min)),
  plot.extent.x.sf = 1,
  plot.extent.y.sf = 1.2,
  x.centre.range = 0.02 * (grid.max - centre.y),
  label.centre.y = FALSE,
  grid.line.width = 0.5,
  gridline.min.linetype = "longdash",
  gridline.n2.linetype = "longdash",
  gridline.n3.linetype = "longdash",
  gridline.n4.linetype = "longdash",
  gridline.n5.linetype = "longdash",
  gridline.max.linetype = "longdash",
  gridline.min.colour = "grey",
  gridline.n2.colour = "grey",
  gridline.n3.colour = "grey",
  gridline.n4.colour = "grey",
  gridline.n5.colour = "grey",
  gridline.max.colour = "grey",
  grid.label.size = 4.5,
  grid.label.color = "#555555",
  gridline.label.offset.x = -0.05 * (grid.max - centre.y),
  gridline.label.offset.y = -0.5,
  label.gridlines.show = c(TRUE, TRUE, TRUE, TRUE, TRUE, TRUE),
  axis.label.offset = 1.15,
  axis.label.size = 5,
  axis.line.colour = "grey",
  group.line.width = 1.5,
  group.point.size = 3,
  group.colours = NULL,
  background.circle.colour = "#D7D6D1",
  background.circle.transparency = 0.2,
  plot.legend = if (nrow(plot.data) > 1) TRUE else FALSE,
  legend.title = "",
  plot.title = "",
  legend.text.size = 14,
  legend.position = "bottom",
  fill = FALSE,
  fill.alpha = 0.3
)
}
\arguments{
\item{plot.data}{dataframe comprising one row per group}

\item{base.size}{text size}

\item{font.radar}{text font family}

\item{values.radar}{values to print at minimum, 'average', and maximum gridlines}

\item{axis.labels}{names of axis labels if other than column names supplied via plot.data}

\item{grid.min}{value at which mininum grid line is plotted}

\item{grid.max}{value at which maximum grid line is plotted}

\item{centre.y}{value of y at centre of plot}

\item{plot.extent.x.sf}{controls relative size of plot horizontally}

\item{plot.extent.y.sf}{controls relative size of plot vertically}

\item{x.centre.range}{controls axis label alignment}

\item{label.centre.y}{whether value of y at centre of plot should be labelled}

\item{grid.line.width}{width of gridline}

\item{gridline.min.linetype}{line type of minimum gridline}

\item{gridline.max.linetype}{line type of maximum gridline}

\item{gridline.min.colour}{colour of minimum gridline}

\item{gridline.max.colour}{colour of maximum gridline}

\item{grid.label.size}{text size of gridline label}

\item{axis.label.offset}{vertical displacement of axis labels from maximum grid line, measured relative to circle diameter}

\item{axis.label.size}{text size of axis label}

\item{axis.line.colour}{colour of axis line}

\item{group.line.width}{line width of group}

\item{group.point.size}{point size of group}

\item{group.colours}{colour of group}

\item{background.circle.colour}{colour of background circle/radar}

\item{background.circle.transparency}{transparency of background circle/radar}

\item{plot.legend}{whether to include a plot legend}

\item{legend.title}{title of legend}

\item{plot.title}{title of radar plot}

\item{legend.text.size}{text size in legend}

\item{legend.position}{position of legend, valid values are "top", "right", "bottom", "left"}

\item{fill}{whether to fill polygons}

\item{fill.alpha}{if filling polygons, transparency values}

\item{grid.mid}{value at which 'average' grid line is plotted}

\item{gridline.mid.linetype}{line type of 'average' gridline}

\item{gridline.mid.colour}{colour of 'average' gridline}

\item{gridline.label.offset}{displacement to left/right of central vertical axis}

\item{label.gridline.min}{whether or not to label the mininum gridline}

\item{label.gridline.mid}{whether or not to label the 'mininum'average' gridline}

\item{label.gridline.max}{whether or not to label the maximum gridline}
}
\value{
a ggplot object
}
\description{
This is ggradar2.r funciton from https://github.com/pedrosdr/ggradar/tree/ggradar2
}
\examples{

library(ggradar2)

# Example 1 (one group):
df <- data.frame(
      NAME = 'G1',
      A = 0.4,
      B = 0.8,
      C = 0.5,
      D = 0.8,
      E = 0.5
      )
ggradar2(df)

# Example 2 (grid line setting):
df <- data.frame(
        NAME = 'G1',
        A = 123,
        B = 65,
        C = 157,
        D = 181,
        E = 99
        )
ggradar2(df,
        # radar values
        values.radar = c("0", "50", "100", "150", "", "200"),
        grid.min = 0,
        grid.n2 = 50,
        grid.n3 = 100,
        grid.n4 = 150,
        grid.max = 200,
        # grid lines
        gridlines.show = c(TRUE, TRUE, TRUE, TRUE, FALSE, TRUE),
        gridline.n2.colour = "blue",
        gridline.n4.colour = "blue",
        # grid labels
        gridline.label.offset.y = 0,
        # axis labels
        axis.label.offset = 1.2
        )

# Example 3 (multiple groups):
df <- data.frame(
        NAME = c('G1', 'G2', 'G3'),
        A = c(0.2, 0.4, 0.5),
        B = c(0.3, 0.6, 0.3),
        C = c(0.7, 0.3, 0.5),
        D = c(0.4, 0.8, 0.2),
        E = c(0.5, 0.6, 0.3)
        )
ggradar2(df)

# Example 4 (fill setting):
df <- data.frame(
        NAME = c('G1', 'G2', 'G3'),
        A = c(2, 4, 5),
        B = c(3, 6, 3),
        C = c(7, 3, 5),
        D = c(4, 8, 2),
        E = c(5, 6, 3)
        )
ggradar2(df,
        values.radar = c("0", "2", "4", "6", "8", "10"),
        grid.min = 0,
        grid.n2 = 2,
        grid.n3 = 4,
        grid.n4 = 6,
        grid.n5 = 8,
        grid.max = 10,
        grid.line.width = 0.3,
        # group parameters
        group.line.width = 0.5,
        group.point.size = 2,
        group.colours = c("red","orange","green"),
        fill = TRUE,
        fill.alpha = 0.5,
        axis.label.offset = 1.15,
        axis.label.size = 4,
        background.circle.colour = "white"
        )

# Example 5 (default ggradar plot): 
df <- data.frame(
        NAME = 'G1',
        A = 0.4,
        B = 0.8,
        C = 0.5,
        D = 0.8,
        E = 0.5
       )
ggradar2(df,
        # radar values
        values.radar = c("0\%", "50\%", "", "", "", "100\%"),
        grid.n2 = 0.5,
        # grid labels
        grid.label.color = "#000000",
        grid.label.size = 6,
        gridline.label.offset.x = -0.1,
        # grid lines
        gridlines.show = c(TRUE, TRUE, FALSE, FALSE, FALSE, TRUE),
        gridline.n2.colour = "#007A87",
        # group
        group.point.size = 6
        )

#End Examples
}
