% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggradar2.R
\name{ggradar2}
\alias{ggradar2}
\title{ggradar2}
\usage{
ggradar2(
  plot.data,
  base.size = 15,
  font.radar = "sans",
  values.radar = c("0\%", "20\%", "40\%", "60\%", "80\%", "100\%"),
  label.gridlines.show = c(TRUE, TRUE, TRUE, TRUE, TRUE, TRUE),
  axis.labels = colnames(plot.data)[-1],
  centre.y = grid.min - ((1/9) * (grid.max - grid.min)),
  plot.extent.x.sf = 1,
  plot.extent.y.sf = 1.2,
  x.centre.range = 0.02 * (grid.max - centre.y),
  label.centre.y = FALSE,
  grid.line.width = 0.5,
  gridline.min.linetype = "longdash",
  gridline.n2.linetype = "longdash",
  gridline.n3.linetype = "longdash",
  gridline.n4.linetype = "longdash",
  gridline.n5.linetype = "longdash",
  gridline.max.linetype = "longdash",
  gridlines.show = c(TRUE, TRUE, TRUE, TRUE, TRUE, TRUE),
  gridline.min.colour = "grey",
  gridline.n2.colour = "grey",
  gridline.n3.colour = "grey",
  gridline.n4.colour = "grey",
  gridline.n5.colour = "grey",
  gridline.max.colour = "grey",
  grid.label.size = 4.5,
  grid.label.color = "#555555",
  gridline.label.offset.x = -0.05 * (grid.max - centre.y),
  gridline.label.offset.y = -0.5,
  axis.label.offset = 1.15,
  axis.label.size = 5,
  axis.line.colour = "grey",
  group.line.width = 1.5,
  group.point.size = 3,
  group.colours = NULL,
  background.circle.colour = "#D7D6D1",
  background.circle.transparency = 0.2,
  plot.legend = if (nrow(plot.data) > 1) TRUE else FALSE,
  legend.title = "",
  plot.title = "",
  legend.text.size = 14,
  legend.position = "bottom",
  fill = FALSE,
  fill.alpha = 0.3
)
}
\arguments{
\item{plot.data}{dataframe comprising one row per group}

\item{base.size}{text size}

\item{font.radar}{text font family}

\item{values.radar}{values to print at 6 intervals for gridlines}

\item{label.gridlines.show}{whether labels of grids should be shown}

\item{axis.labels}{names of axis labels if other than column names supplied via plot.data}

\item{centre.y}{value of y at centre of plot}

\item{plot.extent.x.sf}{controls relative size of plot horizontally}

\item{plot.extent.y.sf}{controls relative size of plot vertically}

\item{x.centre.range}{controls axis label alignment}

\item{label.centre.y}{whether value of y at centre of plot should be labelled, default is FALSE}

\item{grid.line.width}{width of gridline}

\item{gridline.min.linetype}{line type of minimum gridline, default is "longdash". Line types can either be specified as an integer (0=blank, 1=solid, 2=dashed, 3=dotted, 4=dotdash, 5=longdash (default), 6=twodash), or as one of the character strings "blank", "solid", "dashed", "dotted", "dotdash", "longdash", or "twodash", where "blank" uses ‘invisible lines’ (i.e., does not draw them).  The aforementioned settings also apply to gridline.n2.linetype, gridline.n3.linetype, gridline.n4.linetype, gridline.n5.linetype, and gridline.max.linetype.}

\item{gridline.n2.linetype}{line type of No.2 gridline, default is "longdash"}

\item{gridline.n3.linetype}{line type of No.3 gridline, default is "longdash"}

\item{gridline.n4.linetype}{line type of No.4 gridline, default is "longdash"}

\item{gridline.n5.linetype}{line type of No.5 gridline , default is "longdash"}

\item{gridline.max.linetype}{line type of maximum gridline, default is "longdash"}

\item{gridlines.show}{whether gridlines should be shown}

\item{gridline.min.colour}{colour of minimum gridline, default is "grey"}

\item{gridline.n2.colour}{colour of No.2 gridline, default is "grey"}

\item{gridline.n3.colour}{colour of No.3 gridline, default is "grey"}

\item{gridline.n4.colour}{colour of No.4 gridline, default is "grey"}

\item{gridline.n5.colour}{colour of No.5 gridline, default is "grey"}

\item{gridline.max.colour}{colour of maximum gridline, default is "grey"}

\item{grid.label.size}{text size of gridline label, default is 4.5}

\item{grid.label.color}{text colour of gridline label, default is "grey30"}

\item{gridline.label.offset.x}{displacement to left/right of central vertical axis}

\item{gridline.label.offset.y}{displacement to top/bottom of central vertical axis}

\item{axis.label.offset}{vertical displacement of axis labels from maximum grid line, measured relative to circle diameter, default is 1.15}

\item{axis.label.size}{text size of axis label, default is 5}

\item{axis.line.colour}{colour of axis line, default is "grey"}

\item{group.line.width}{line width of group, default is 1}

\item{group.point.size}{point size of group, default is 2}

\item{group.colours}{colour of group, default is NULL}

\item{background.circle.colour}{colour of background circle/radar, default is "grey90"}

\item{background.circle.transparency}{transparency of background circle/radar, default is 0.2}

\item{plot.legend}{whether to include a plot legend}

\item{legend.title}{title of legend}

\item{plot.title}{title of radar plot}

\item{legend.text.size}{text size in legend, default is 14}

\item{legend.position}{position of legend, valid values are "top", "right", "bottom", "left"}

\item{fill}{whether to fill polygons, TRUE or FALSE? Default is FALSE}

\item{fill.alpha}{if filling polygons, transparency values, default is 0.3}
}
\value{
a ggplot object
}
\description{
This is ggradar2.r function from https://github.com/pedrosdr/ggradar/tree/ggradar2, with some minor adjustments made.
}
\examples{

library(ggradar2)

# Example 1 (one group):
df <- data.frame(
      NAME = 'G1',
      A = 0.4,
      B = 0.8,
      C = 0.5,
      D = 0.8,
      E = 0.5
      )
ggradar2(df)

# Example 2 (grid line setting):
df <- data.frame(
        NAME = 'G1',
        A = 123,
        B = 65,
        C = 157,
        D = 181,
        E = 99
        )
ggradar2(df,
        # radar values
        values.radar = c("0", "50", "100", "150", "", "200"),
        # grid lines
        gridline.n2.colour = "blue",
        gridline.n4.colour = "blue",
        # grid labels
        gridline.label.offset.y = 0,
        # axis labels
        axis.label.offset = 1.2
        )

# Example 3 (multiple groups):
df <- data.frame(
        NAME = c('G1', 'G2', 'G3'),
        A = c(0.2, 0.4, 0.5),
        B = c(0.3, 0.6, 0.3),
        C = c(0.7, 0.3, 0.5),
        D = c(0.4, 0.8, 0.2),
        E = c(0.5, 0.6, 0.3)
        )
ggradar2(df)

# Example 4 (fill setting):
df <- data.frame(
        NAME = c('G1', 'G2', 'G3'),
        A = c(2, 4, 5),
        B = c(3, 6, 3),
        C = c(7, 3, 5),
        D = c(4, 8, 2),
        E = c(5, 6, 3)
        )
ggradar2(df,
        values.radar = c("0", "2", "4", "6", "8", "10"),
        grid.line.width = 0.3,
        # group parameters
        group.line.width = 0.5,
        group.point.size = 2,
        group.colours = c("red","orange","green"),
        fill = TRUE,
        fill.alpha = 0.2,
        axis.label.offset = 1.15,
        axis.label.size = 4,
        background.circle.colour = "white"
        )

# Example 5 (default ggradar plot): 
df <- data.frame(
        NAME = 'G1',
        A = 0.4,
        B = 0.8,
        C = 0.5,
        D = 0.8,
        E = 0.5
       )
ggradar2(df,
        # radar values
        values.radar = c("0\%", "50\%", "", "", "", "100\%"),
        # grid labels
        grid.label.color = "#000000",
        grid.label.size = 6,
        gridline.label.offset.x = -0.1,
        gridline.n2.colour = "#007A87",
        # group
        group.point.size = 6
        )

#End Examples
}
